[#6677] Remove rule_texts_for_tests.py
Begone, vile beast.
